[batch] Extract phase calculation types and time ordering function to a new module
